Fix for 64 bit systems.
git-svn-id: https://ravl.svn.sourceforge.net/svnroot/ravl/main/Ravl@1357 7226f50c-d813-0410-99bf-979b288ab488

+++ 3D/Mesh/TriMeshBinIO.cc
@@ -36,7 +36,7 @@ namespace Ravl3DN {
     s << ts.Faces().Size(); 
     const VertexC *x = &(ts.Vertices()[0]);
     for(SArray1dIterC<TriC> it(ts.Faces());it;it++) {
-      s << (it->VertexPtr(0) - x)  << (it->VertexPtr(1) - x) << (it->VertexPtr(2) - x);
+      s << (int) (it->VertexPtr(0) - x)  << (int) (it->VertexPtr(1) - x) << (int) (it->VertexPtr(2) - x);
       s << it->TextureCoords();
       s << it->Colour();
     }
